Rename sidebar to properties, display aggregate props
