+++ src/assets/main.css
@@ -1,3 +1,6 @@
+ul { padding-left: 0; }
+li { list-style: none; }
+
 .content {
   display: flex;
   flex-flow: row wrap;
