+++ src/ui/data/Observations.jsx
@@ -1,13 +1,13 @@
 'use strict'
 const React = require('react')
 const Map = require('./Map.jsx')
-const Sidebar = require('./Sidebar.jsx')
+const Properties = require('./Properties.jsx')
 
 class Observations extends React.Component {
   render () {
     return (
       <div className='content'>
-        <Sidebar />
+        <Properties />
         <Map />
       </div>
     )
